+++ src/lib/config.ts
@@ -54,8 +54,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.1.1",
-  sdkVersion: "0.2.4-alpha.5",
-  genVersion: "2.548.1",
+  sdkVersion: "0.2.4-alpha.6",
+  genVersion: "2.548.6",
   userAgent:
-    "speakeasy-sdk/typescript 0.2.4-alpha.5 2.548.1 0.1.1 @inkeep/inkeep-analytics",
+    "speakeasy-sdk/typescript 0.2.4-alpha.6 2.548.6 0.1.1 @inkeep/inkeep-analytics",
 } as const;