+++ RELEASES.md
@@ -99,3 +99,13 @@ Based on:
 - [typescript v0.2.4-alpha.5] .
 ### Releases
 - [NPM v0.2.4-alpha.5] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.5 - .
+
+## 2025-03-14 17:52:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.3 (2.548.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.4-alpha.6] .
+### Releases
+- [NPM v0.2.4-alpha.6] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.6 - .